+++ app/src/main/java/com/openclassrooms/realestatemanager/navigation/NavigationScreen.kt
@@ -14,7 +14,7 @@ sealed class NavigationScreen(val route: String) {
     data object Mortgage: NavigationScreen("mortgage_screen") {
         const val ARG_PRICE = "price"
         val routeWithArgs = "$route/{$ARG_PRICE}"
-        fun createRoute(price: Double): String = "$route/$price"
+        fun createRoute(price: Int): String = "$route/$price"
     }
 
     //Form
